deleteUser 수정 password 추가에 따른 삭제 검증 추가

+++ src/main/java/com/example/schedule_jpa/service/UserService.java
@@ -49,8 +49,13 @@ public class UserService {
         return new UserResponseDto(findUser.getId(), findUser.getUserName(), findUser.getEmail());
     }
 
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id, String password) {
         User findUser = userRepository.findByIdOrElseThrow(id);
+
+        if(!findUser.getPassword().equals(password)){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
+
         userRepository.delete(findUser);
     }
 }
